requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
streamlit
torch
torchvision
opencv-python
scikit-image
scipy
gdown
joblib
